receipt-watcher/pyproject.toml
[project]
name = "receipt-watcher"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "stack-shared",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "pyyaml>=6.0",
    "imap-tools>=1.7.0",
    "beautifulsoup4>=4.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
stack-shared = { path = "../shared", editable = true }
